+++ CHANGELOG.md
@@ -32,6 +32,10 @@ BUG FIXES:
 * Updated config_schema.json to include missing values. ([#4712](https://github.com/microsoft/AzureTRE/issues/4712))([#4714](https://github.com/microsoft/AzureTRE/issues/4714))
 * Remove workspace upgrade step from databricks template ([#4726](https://github.com/microsoft/AzureTRE/pull/4726))
 * Update Starlette and FastAPI versions ([#4738](https://github.com/microsoft/AzureTRE/pull/4738))
+* Fix resource processor status updates not being received by API due to session_id mismatch in Service Bus messages
+
+COMPONENTS:
+* Resource Processor updated to 0.13.1
 
 ## 0.25.0 (July 18, 2025)
 **IMPORTANT**: